requirements.txt
djangorestframework
xhtml2pdf